+++ app/schemas/dashboard.py
@@ -1,10 +1,11 @@
 from pydantic import BaseModel
+from decimal import Decimal
 from app.schemas.sale import SaleResponse, FollowUpResponse
 
 
 class DashboardSummary(BaseModel):
     totalCustomers: int
-    salesThisMonth: float
+    salesThisMonth: Decimal
     pendingFollowUps: int
     upcomingPurchases7Days: int
     recentSales: list[SaleResponse] = []
